stage1_scraper/requirements.txt
pyppeteer==1.0.2
python-dotenv==1.0.1
pandas==2.2.1 